newnewsapp/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-crispy-forms==1.11.1
django-extensions==3.1.1
django-filter==2.4.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
entrypoints==0.3
Faker==5.6.1
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
jedi==0.17.0
jieba==0.42.1
jieba3k==0.35.1
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.10
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
mariadb==1.0.5
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
news-api==1.1.0
newsapi==0.1.1
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.5
notebook==6.1.6
numpy==1.19.5
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.10
pycparser==2.20
Pygments==2.7.3
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
snownlp==0.12.3
soupsieve==2.1
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
true-encoding==1.0.2
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
zope.interface==5.2.0
